Webview switcher patch
Mulch and the geuinie LineageOS webview use the same module name, which
causes build errors. Both repositories however are needed - the default
build uses the LineageOS webview, the microG builds use Mulch.

To solve the conflict, a patch is applied to the respective "unused"
repository to rename the module to an 'unused' one.

Change-Id: I2d0ed8152f569195b693a6db09a7460e100bb609